+++ bot/twitter/src/main/java/ReceivePubSubMessage.java
@@ -1,0 +1,49 @@
+import java.util.logging.Logger;
+
+import org.apache.beam.sdk.Pipeline;
+import org.apache.beam.sdk.io.gcp.pubsub.PubsubIO;
+import org.apache.beam.sdk.io.gcp.pubsub.PubsubMessage;
+import org.apache.beam.sdk.transforms.DoFn;
+import org.apache.beam.sdk.transforms.ParDo;
+
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+public class ReceivePubSubMessage {
+
+	public static final Logger LOGGER = Logger.getAnonymousLogger();
+
+	private static class SampleDoFn extends DoFn<PubsubMessage, String> {
+		@ProcessElement
+		public void process(ProcessContext c) {
+
+			// レガシーサイトのURLを取得する
+			String url = new String(c.element().getPayload());
+			LOGGER.info(String.format("Receive Message : %s", url));
+			c.output(url.toUpperCase());
+
+			// レガシーサイトの判定結果を取得する
+			String attribute = new String(c.element().getAttribute("siteType"));
+			LOGGER.info(String.format("Receive Attribute : %s", attribute));
+			c.output(attribute.toUpperCase());
+
+			if (attribute.equals("old")) {
+				TwitterBot twitterBot = new TwitterBot();
+				twitterBot.postToTwitter(url);
+			}
+		}
+	}
+
+	public static void main(String[] args) {
+		Pipeline p = Pipeline.create();
+
+		// input
+		p.apply(PubsubIO.readMessagesWithAttributes()
+				.fromSubscription("{SUBSCRIPTION_NAME}"))
+				// output
+				.apply("Sample pardo", ParDo.of(new SampleDoFn()));
+
+		p.run();
+	}
+
+}
